pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "triplequant-vlm"
version = "0.1.0"
description = "Production-grade quantization and benchmarking for vision-language models"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }

dependencies = [
    # Core ML
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "transformers>=4.46.0",
    "accelerate>=0.33.0",

    # Quantization
    "llmcompressor>=0.4.0",

    # Inference / benchmarking
    "vllm>=0.5.0",

    # Datasets
    "datasets>=2.20.0",
    "Pillow>=10.0.0",

    # Config validation
    "pydantic>=2.7.0",
    "pyyaml>=6.0",

    # Utilities
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
tq-quantize  = "quantize:main"
tq-benchmark = "benchmark:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
